+++ src/DefaultCodes.ts
@@ -1,3 +1,6 @@
 import ErrorCode from "./ErrorCode";
 
-export const CODE_MISC_INTERNAL_ERROR = new ErrorCode("misc/internal-error", `An internal error was caught but the operation couldn't be completed`);
+export const ERROR_CODE_MISC_INTERNAL_ERROR = new ErrorCode({
+  id: "misc/internal-error",
+  defaultMessage: `An internal error occurred`,
+});