+++ CHANGELOG.md
@@ -2,10 +2,11 @@
 
 ## Breaking Changes
 
+- remove builder API
 - strip API back to basics -- complex features will come back over time
 - remove visibility/scoped value API for now
 - `messages` -> `message`
-- remove builder properties from `SendableError`. If you want to edit an error use `SendableError.of(oldError).message("...")`
+- remove builder properties from `SendableError`. If you want to edit an error use `SendableError.of(oldError, { message: "...") }`
 
 # 0.0.x